src/main/java/nextstep/courses/domain/Course.java
package nextstep.courses.domain;

import nextstep.courses.domain.course.Sessions;

import java.time.LocalDateTime;
import java.util.Objects;

public class Course {
    private Long id;

    private String title;

    private Sessions sessions = new Sessions();

    private Long creatorId;

    private LocalDateTime createdAt;

    private LocalDateTime updatedAt;


    public Course() {
    }

    public Course(String title, Long creatorId) {
        this(0L, title, creatorId, LocalDateTime.now(), null);
    }

    public Course(Long id, String title, Long creatorId, LocalDateTime createdAt, LocalDateTime updatedAt) {
        this.id = id;
        this.title = title;
        this.creatorId = creatorId;
        this.createdAt = createdAt;
        this.updatedAt = updatedAt;
    }
    public long getId() {
        return id;
    }

    public String getTitle() {
        return title;
    }

    public Long getCreatorId() {
        return creatorId;
    }

    public LocalDateTime getCreatedAt() {
        return createdAt;
    }

    public void addSession(Session session) {
        this.sessions.add(session);
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Course course = (Course) o;
        return Objects.equals(id, course.id) && Objects.equals(title, course.title) && Objects.equals(sessions, course.sessions) && Objects.equals(creatorId, course.creatorId);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, title, sessions, creatorId);
    }

    @Override
    public String toString() {
        return "Course{" +
                "id=" + id +
                ", title='" + title + '\'' +
                ", creatorId=" + creatorId +
                ", createdAt=" + createdAt +
                ", updatedAt=" + updatedAt +
                '}';
    }
}
